Remove role-related functionality and simplify user management
- Remove RoleEntity, RoleDto, and related classes
- Remove roles field from UserEntity and UserDto
- Remove role-related operations from UserService
- Update UserMapper to remove role mapping
- Remove RoleNotFoundException and SaveUserRoleException from GlobalExceptionHandler
- Update database schema to remove roles and users_roles tables
- Add @NonNull annotations to AddressDto fields
- Remove deleteByUserId method from AddressRepository and its implementation
- Update AddressService to simplify update logic
- Update UserJdbcRepositoryImpl to remove role-related queries
- Move validation classes to appropriate packages
