+++ src/main/java/app/jaba/dtos/AddressDto.java
@@ -2,6 +2,7 @@ package app.jaba.dtos;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import io.swagger.v3.oas.annotations.media.Schema;
+import org.springframework.lang.NonNull;
 
 import java.io.Serializable;
 import java.util.UUID;
@@ -12,18 +13,23 @@ public record AddressDto(
         @Schema(description = "Unique identifier of the address", example = "123e4567-e89b-12d3-a456-426614174000", readOnly = true)
         UUID id,
 
+        @NonNull
         @Schema(description = "Street name", example = "Rua das Flores")
         String street,
 
+        @NonNull
         @Schema(description = "City name", example = "Rio de Janeiro")
         String city,
 
+        @NonNull
         @Schema(description = "State name", example = "RJ")
         String state,
 
+        @NonNull
         @Schema(description = "ZIP code", example = "50712-432")
         String zip,
 
+        @NonNull
         @Schema(description = "House number", example = "13")
         String number
 ) implements Serializable {
