--- src/main/java/app/jaba/dtos/RoleDto.java
@@ -1,5 +1,0 @@
-package app.jaba.dtos;
-
-public enum RoleDto {
-    OWNER, CLIENT
-}
