--- src/main/java/app/jaba/entities/RoleEntity.java
@@ -1,19 +1,0 @@
-package app.jaba.entities;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.FieldDefaults;
-
-import java.util.UUID;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class RoleEntity {
-
-    UUID id;
-    String name;
-}
