--- src/main/java/app/jaba/entities/UserRoleEntity.java
@@ -1,21 +1,0 @@
-package app.jaba.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.FieldDefaults;
-
-import java.util.UUID;
-
-import static lombok.AccessLevel.PRIVATE;
-
-@Data
-@FieldDefaults(level = PRIVATE)
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class UserRoleEntity {
-    UUID userId;
-    UUID roleId;
-}
