--- src/main/java/app/jaba/exceptions/RoleNotFoundException.java
@@ -1,7 +1,0 @@
-package app.jaba.exceptions;
-
-public class RoleNotFoundException extends RuntimeException {
-    public RoleNotFoundException(String message) {
-        super(message);
-    }
-}
