--- src/main/java/app/jaba/exceptions/SaveUserRoleException.java
@@ -1,7 +1,0 @@
-package app.jaba.exceptions;
-
-public class SaveUserRoleException extends RuntimeException {
-    public SaveUserRoleException(String message) {
-        super(message);
-    }
-}
