+++ src/main/java/app/jaba/mappers/UserMapper.java
@@ -7,7 +7,7 @@ import app.jaba.entities.UserEntity;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
-@Mapper(componentModel = "spring", uses = {AddressMapper.class, RoleMapperImpl.class})
+@Mapper(componentModel = "spring", uses = {AddressMapper.class})
 public interface UserMapper {
     @Mapping(target = "id", ignore = true)
     @Mapping(target = "lastUpdate", ignore = true)
