+++ src/main/java/app/jaba/repositories/AddressJdbcRepositoryImpl.java
@@ -92,10 +92,4 @@ public class AddressJdbcRepositoryImpl implements AddressRepository {
                 .optional();
     }
 
-    @Override
-    public void deleteByUserId(UUID userId) {
-        jdbcClient.sql("DELETE FROM addresses WHERE user_id = :user_id")
-                .param("user_id", userId)
-                .update();
-    }
 }
