--- src/main/java/app/jaba/repositories/RoleRepository.java
@@ -1,17 +1,0 @@
-package app.jaba.repositories;
-
-import app.jaba.entities.RoleEntity;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface RoleRepository {
-
-    Optional<RoleEntity> findById(UUID id);
-
-    List<RoleEntity> findAll();
-
-    Optional<RoleEntity> findByName(String name);
-
-}
