--- src/main/java/app/jaba/services/validations/CreateUserRoleValidation.java
@@ -1,7 +1,0 @@
-package app.jaba.services.validations;
-
-import app.jaba.entities.UserRoleEntity;
-
-public interface CreateUserRoleValidation {
-    void validate(UserRoleEntity userRoleEntity);
-}
