+++ src/main/java/app/jaba/services/validations/user/CheckLoginUpdateValidation.java
@@ -1,28 +1,21 @@
-package app.jaba.services.validations;
+package app.jaba.services.validations.user;
 
 import app.jaba.entities.UserEntity;
-import app.jaba.exceptions.LoginAlreadyInUseException;
 import app.jaba.repositories.UserRepository;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Component;
 
-import java.util.Objects;
-
 @Component
 @RequiredArgsConstructor
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
-public class LoginAlreadyInUseValidation implements CreateUserValidation {
+public class CheckLoginUpdateValidation implements UpdateUserValidation {
 
     UserRepository userRepository;
 
     @Override
     public void validate(UserEntity user) {
-        userRepository.findByLogin(user.getLogin()).ifPresent(u -> {
-            if (!Objects.equals(u.getId(), user.getId())) {
-                throw new LoginAlreadyInUseException("Login already in use");
-            }
-        });
+        // TODO: VALIDAR SE O LOGIN É O MESMO DURANTE ATUALIZACAO E NAO PODE SER DIFERENTE
     }
 }
