+++ src/main/resources/db/migration/V1__init_schema.sql
@@ -7,19 +7,6 @@ CREATE TABLE IF NOT EXISTS users (
     last_update TIMESTAMP DEFAULT NOW()
 );
 
-CREATE TABLE IF NOT EXISTS roles (
-    id UUID PRIMARY KEY,
-    name VARCHAR(255) NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS users_roles (
-    user_id UUID NOT NULL,
-    role_id UUID NOT NULL,
-    PRIMARY KEY (user_id, role_id),
-    FOREIGN KEY (user_id) REFERENCES users (id),
-    FOREIGN KEY (role_id) REFERENCES roles (id)
-);
-
 CREATE TABLE IF NOT EXISTS addresses (
     id UUID PRIMARY KEY,
     street VARCHAR(255),
